changed to kind

+++ driver2.f90
@@ -25,8 +25,8 @@ program driver2
   use wizard
   implicit none
   integer :: j
-  double precision :: rhotot
-  double precision, allocatable :: ccsf(:), ddsf(:)
+  real(kind=dp) :: rhotot
+  real(kind=dp), allocatable :: ccsf(:), ddsf(:)
 
   verbose = 1
 
@@ -35,24 +35,24 @@ program driver2
 
   call initialise
 
-  lb = 120.0
-  arep = 25.0
-  arep(1, 2) = 30.0
-  z(1) = 1
-  z(2) = -1
+  lb = 120.0_dp
+  arep = 25.0_dp
+  arep(1, 2) = 30.0_dp
+  z(1) = 1.0_dp
+  z(2) = -1.0_dp
 
   call dpd_potential
 
-  rhotot = 3.0d0
+  rhotot = 3.0_dp
 
-  rho(1) = 0.5d0 * rhotot
+  rho(1) = 0.5_dp * rhotot
   rho(2) = rho(1)
 
   call write_params
 
   call hnc_solve
 
-  if (error .gt. 1.0d-10) &
+  if (error .gt. 1.0E-10_dp) &
        & print *, 'Warning, did not converge to 1e-10'
 
   allocate(ccsf(ng-1))
